Implement named FunctionExpressions; fix function call scoping bug
It turns out that a function expression can have an identifier; in the
body of

    f = function foo() { /* … */ };

f and foo both refer to the function.  (This now works.)

Also discovered and fixed a bug where we did not use the saved scope
when invoking a function.  This basically meant that closures were
totally broken; an unbound variable in the function body would refer to
the slot found in the caller’s scope, not in the scope at the time the
function expression was evaluated.  (This hadn’t shown up before
because we always evaluated the function in the same scope as it was
defined.)