Implement unflattening of pointer values
Note that there are two versions of this code, one broken (for circular
data) and one correct.  For now the correct one is commented out,
because I want to be sure to write a test to catch the broken version,
but it’s not possible to write that test yet because we need interfaces
or structs.