Fix playground
Make interpreter.js, serialize.js and autoexec.js compatible with both
Node.js and browsers.

In the browser each of these just defines various global variables;
ones which are “exports” are initialised, while “imports” are left
uninitialised and will be initialised in another file.

In node, those “global” var declarations are actually module-local, but
an epilogue uses require to initialise the imports and adds the
exported ones to module.exports.