Fix eval scoping issues
Previously eval behaved as it would in non-strict mode—i.e., variables
declared in the evaled code were added to the outer scope.

(This is probably still not *quite* spec-correct: in the case of a
renamed eval (e.g., `myEval = eval; myEval('var ...`) it should behave
as if non-strict even when called from strict mode code.  Maybe we will
choose not to implement that particular detail.)