Fix issue with CatchClause
The compiler caught an issue with CatchClause: if there was no param
given, the body would get executed with no scope (none at all).

Fortunately the catch clause param is not optional, so to fix the
compiler warning we need merely remove the if statement surrounding the
code that sets up the inner scope.