Have .schedule() return start time for next sleeping thread
Previously it just returned a boolean indicating whether there was a
READY (runnable) thread.  Now it return 0 in that case or an absolute
time at which to resume the next sleeping thread otherwise.

Also fix bug with async function callbacks attempting to wake the
current thread rather than (necessarily) the thread which is BLOCKED.