Fix race condition in testConnectionListenUnlisten
In very rare circumstances the running interpreter may not have
finished opening a socket to listen on before the sideFunc is run.  In
much more common circumstances a similar thing happens with a
not-yet-committed test of connectionWrite.

In both cases the fix is to have the test client function run as a
native function from within the interpreter only once connectionListen
has completed.

It seems that the sideFunc functionality of runAsyncTest is probably
not to be useful.  Consider removing this if no other tests end up
using it.