Replace .running with .status
And have three statuses instead of two, to make room for a mechanism do
do timer and network socket resurrection.

Also fix a bug which could result in a timeout causing a spurious call
to the repeat inner function of .go_ even after a previous invocation
of repeat had determined that there were no outstanding (user)
timeouts.  This would result in an unhandled exception if the
interpreter's status had in the mean time been set to STOPPED.