Have nginx serve 503 error page if codecity server down
This change configures nginx to serve 503 Unavailable if it would
otherwise server a 502 Bad Gateway or 504 Gateway Timeout, and returns
a static error page with the logo, tagline and a message asking the
user to try again later.

The error page is cribbed from $.pages.home.jssp; it would be nice if
the 503 error page could be served by the same code (for future
stylistic consistency), but of course that is impossible.  They will
have to be kept in sync manually.  :-(