Add nginx config for local development
This is based on my local copy, which in turn is based on the canonical
example cc.conf.  Only the path to the static directory needs to be
configured.

Now that we have HTTP/1.0, we no longer need to listen on port 80 to
work around Chrome's restrictions against HTTP/0.9, so this is
configured to listen on :8080.