Throw first, ask questions later
When throwing a (user) exception -
 pseudo-Error or otherwise - throw the exception value and then unwind the stack in step() or run(), instead of unwinding the stack and then throwing STEP_ERROR.

This is cleaner, allows native functions to catch internally-generated errors if desired, and ensures we never try to unwind the wrong stack.

It *will* cause the interpreter to blow up if an async native function tries to throw an error, but we want to know about that because it's a bug which would previously have resulted in possilby the wrong stack being unwound.