Introduce ObjectLike, Box, and ToObject
In many places in the ES spec, the ToObject function is used to denote that a Boolean, Number or String object will be created from the corresponding primitive value.  (This is used e.g. when resolving "foo".length.)

We don't support boxed primitives, and there are few situations where a strict-mode program would create a boxed primitive without an explicit call to one of those constructors, but implementation of many functions is easier if we have temporary boxed primitives.  Box is such a temporary box.

To ensure that Boxes are never user visible, we do not make Box @extend Interpreter.prototoype.Object, but instead both it and ...Object implement the new Interpreter.ObjectLike interface which will be used to type temprorary variables internally.