Better type checking for NativeFunctions
It turns out that although closure-compiler was able to recognise that the call: and construct:  implementations conformed to the required type signature, it wasn't actually fully type-checking usage of the parameters inside the body fo the implementation.  Explicitly declaring the types of these functions fixes that, forces a few additional explicit casts, and brings us to over 92% typed.