Make Array.prototype.toString non-generic
Since intrp.Array.prototype.toString is now no longer used as the implementation of the built-in of the same name, it's no longer possible for users to apply() it to non-intrp.Arrays, and so it need not be generic.

Therefore, it has been rewritten to remove code that was only there to make it generic, and also modified to ensure that it uses the intrp.Object API so that property access restrictions will be enforced.