Mark non-constructability with null
Previously non-constructable functions has .construct set to a function which would throw TypeError with a suitable message.

This creates a problem when we wish to determine whether a function is a constructor or not (without calling its .construct method).  One possible fix is just to have a single implementation of the TypeError-throwing .construct, and use it as a sentinel value:

    if (f.construct === intrp.nonConstructableConstructImplementation) {
      // f can't be constructed!
    }

But simpler approach, which more closely parallels the spec, is to have .construct be null for non-constructors.

It is not necessary to do the same for .call, since (as far as I am able to determine) there are NO JS function objects which are not callable: class constructors like WeakMap or class Foo{} may throw TypeError if you try to call them without 'new', but IsCallable still returns true for them, as can be demonstrated using Array.prototype.toString:

    // WeakMap is callable (but throws TypeError):
    console.log('null:', Array.prototype.toString.apply({join: WeakMap}));
    // WeakMap is callable (but throws TypeError):
    console.log('null:', Array.prototype.toString.apply({join: WeakMap}));

N.B. I've added a full listings of the @params on UserFunction.prototype.construct, which is also marked @override.  This is arguably a violation of the styleguide, but since Interpreter.prototype.construct now has its type declared in a way which provides no documentation about the meaning of the parameters it seemed helpful to make sure they are documented *somewhere*.