Introduce OldConstructorFunction
OldNativeFunction instances could preiviously be callable or not callable, but the difference depended on the value of their .illegalConstructor property, rather than on whether or not .construct was null.

OldConstructorFunction is a new subclass of OldNativeFunction that is used for all old-style native functions which are constructors, and is bequeathed the .construct implementation from OldNativeFunction, which is now used  (other than as a base class for this new class) for only non-constructible old-style native functions.