Avoid instanceof
There are quite a few places in the interpreter where we do

    value instanceof intrp.Object

when we could do

    typeof value === 'object' && value !== null

The latter ought in principle to be faster, because it doesn't require traversing value's prototype chain.

This commit tests whether this is true.  The results are not totally conclusive, but it appears the answer is: is is not.  Using instanceof is at least no slower, and is probably marginally faster.