Serialize using property descriptors
Previously we would serialize property values and attributes seperately; values would be stored in .props, while attributes would be in .nonWritable, .nonEnumerable and .nonConfigurable (with absence of a property name denoting that the respective attribute was true).

Now we store the whole property descriptor for each property, in the same format used by Object.defineProperty (except that objects are of course replaced with their reference nubmers).