Restrict new hack to string literals
There's a risk that users will do:

    var foo = ...;
    var obj = new foo;

when foo is supposed to be a constructor but unexpectedly contains a string primitive, and consequently receive a builtin (or confusing error message) instead of the new object they were expecting.

To prevent this likely-confusing scenario, restrict use of the new hack to string literals (with eval as a workaround for the rare cases where one wants to compute the value of the builtin to obtain).