Introduce Interpreter.Source
The Source class allows us to abstract away details of how source code is stored—e.g., when storing a function' source, do we keep the whole original source text, or just the substring containing the function declaration/expression?

Initial implementation stores the whole thing, but it would be easy to change this in future if we want to.