Don't store UserFunction source twice
Previously we stored the source code for UserFunctions twice: we stored the whole source file in the AST and the source for just the function itself on the function.

Now that we store it as Source objects instead of strings, and so both are the same (literally the same object), there is no reason to store the second copy on the function object.