Verify callee is a Function earlier
In CallExpression, verify that callee evaluated to an intpr.Funciton before the end of step 2, so that if the function's .call or .constrcut method returns FunctionResult.CallAgain we don't bother doing this check again.

Also: store the evaluated callee value in state.tmp_ until it is verified to be a function.

Re-name and CallInfo.callee to .func and give it @type {?Interpreter.prototype.Function}.  Since the spec requires we evaluate arguments before testing type of callee, .func will need to be some non-intrp.Function type, and null seems like as good a choice as any.