Re-use State and CallInfo for .apply, .call and BoundFunction invocation
Have built-in F.p.apply and F.p.call write the new func/this/args value into state.info_, so that subsequent invocations the CallExpression step function will short-circuit directly to the function being called/applied.

This improves efficiency when calling functions that might return FunctionResult.CallAgain, and avoids a heretofore-unnoticed potential issue where such a function would, if invoked using .apply, on each subsequent invocation receive a different args array (albeit with the same actual argument values in it).

Relatedly, make BoundFunction.p.call and BF.p.construct also rewrite the CallInfo structure in the same way.

From a stack-trace perpective, we lose information about exaclty how the end function was invoked (i.e., we don't know it was via call/apply/bind), but in return we definitely know which function was ultimately invoked, which is rather more important.

Note that this creates a small issue re: callerPerms as observed in the target function, which is documented accordingly.