Move NativeCallImpl, NativeConstructImpl
Since these two typedefs are only used for type checking, and they are used in conjuction with NativeFunctionOptions, move them to be with that declaration (in the 'types not used for storing state' section).

Also: remove the "Interpreter." prefex from these types.  This may turn out to be the wrong thing to do, but for the moment they are not used outside this file, and though they might be in future the same is true of NativeFunctionOptions and indeed all the other types defined in this section.  So be consistent until there is reason not to be.