Don't set intrp.Function.prototype.{call,construct}.id
Due to poor design choices in the intrp.NativeFunction constructor, the default intrp.Function call and construct implementations were getting (incorrectly) assigned .ids of "WeakMap" and "Function.prototype [[construct]]", respectively.

These labels are sort of true (Function.prototype can't be constructed, and WeakMap can't be called, so they both default to these always-throw methods), but potentially misleading and not needed by the serialiser, which never visits intrp.Function.prototype, so the  NativeFunction constructor has been revised not to scribble where it was not intended to.