Factor makeStack from intrp.Error constructor
Rather than trying to create a .stack trace at the point an Error object is created, separate out this code into intrp.Error.prototype.makeStack.

This function is called from the built-in Error constructors (so errors created from userland code are born with a .stack property as before) and from the try/catch in step() and run(), so that errors thrown by any of the "machinery" of the interpreter (e.g., from any NativeFunction, or any of the API methods on intrp.Object) will get a .stack property - with the correct thread's stack trace! - set before thread.stateStack_ is unwound.

There is one small issue with this approach, which is documented in the JSDoc for E.p.makeStack.