Kill correct thread
Turns out that, when called as a result of an error being thrown in an async function, unwind_ will unwind the correct thread's stack but then, if that stack is empty, kill whatever thread intrp.thread happens to be pointing at.

Fix this and add tests to make sure it stays fixed.