Fix infinite loop in intrp.Function.prototype.toString
If f is a NativeFunction (or any other kind of intrp.Function that doesn't override i.F.p.toString), and a user sets f.name to f (or f.name === g and g.name === f, or any other cyclic arrangement) then the *internal* stringification of f (via i.F.p.toString) will get caught in an infinite loop.

To avoid this, only use .name as part of the stringificaiton if it's already a string (as it would normally be).