Check Object coercibility first
The spec says, In MemberExpression, that CheckObjectCoercible / RequireObjectCoercible should be called on baseValue before calling ToPropertyKey (i.e., ToString) on propertyNameValue.

It's not presently possible to write a test for this, but it will become possible should any of the internal toString methods become capable of throwing an error (or should we eventually call userland toString functions).