Look up step function at State creation
We know from painful past experience that looking up step functions is slow—and while it is not nearly as slow as it used to be, a little profiling suggested that it's still the most expensive operation the inerpreter performs.

This is borne out by a ~15% performance improvment by the simple expedient of looking the step function up once at State creation time.