Compile server + tests in one run
Have the tests require the server, so that both can be compiled in a single run.  This speeds up comilation because the comipler, being a Java app, takes much longer to start up than it takes to compile all our code (and there is in any case a very large overlap between the server and tests).

It does slow tests down very, very slightly -  but only by the time taken for V8 to compile codecity.js and the currently zero other modules it requires but the tests don't.