Actually compile tests
It turns out that although the changes made in PR #200 resulted in run.js and a few other files being compiled, closure-compiler was ignoring all the _test.js and _bench.js files because it only looks at require calls which have a string literal argument, whereas we loop through all the files in the directory.

So we need to hard-code a list of files(/modules) to compile somewhere for the compiler.  Yet dynamically running all _test and _bench files is useful to ensure none are overlooked.

So now we have a hard-coded list for compilation, and dynamic loading for running, and additionally runTests and runBenchmarks will warn if the file they are running wasn't on the hard-coded compilation list.