Create code module, quote function
The code module contains functions for handling JS source code that are needed by multiple (not-yet-committed) other modules, including the dumper and selector modules.

The quote function takes a string and converts it to a string literal, choosing to use single or double quotes according to which will be most readable (i.e., require fewest escaped characters).