Function .name auto-setting fixes
Where functions are represented in the output as function expressions, they often get their .name property set automatically—either to the function's identifier, or to the name of the variable or property to which it is assigned.  But this doesn't happen in all cases, so this commit modifies Dumper.prototype.exprForFunction and ObjectInfo.prototype.dumpProperty_ so that their internal model of when .name gets set automatically matches what the pristine interpreter will actually do, so that .name can be set explicity only if required.