Have ObjectInfo.p.dumpRecursively delete unwanted properties
Properties that are created automatically (e.g., Function .name and Error .stack) could be deleted.  Issue delete statements to remove them if they are not present in intrp.

Note that it turns out we don't need to delete Error .message properties, because we can prevent them from being created by not supplying an argument to the constructor.