Use exprForBuiltin when calling Object.create
A bug was found, wherein Object.create would be called before that variable/property existed, because the function name was hardcoded in exprForObject.

This is now fixed.  In the process, it was observed that, as  Object.create is a polyfill, exprForObject will never call the polyfill but always call the builtin via the new hack.