Perform survey to find non-global scopes
Introduce ScopeDumper.prototype.survey and ObjectDumper.prototype.survey methods that recursively visit all objects accessible via the global scope.

The objective is to construct, for each scope, a set of UserFunctions that have that scope as their .scope, and a set of Scopes which have that scope as their .outerScope.

This will allow us to visit all these scopes (and construct these functions in them) when dumping.