Have survey record scopes for Arguments objects
It's possible to encounter arguments objects outisde of the Scope they were created for—e.g.:

var argsObj = (function() {return arguments;})(/* ... */);

(and indeed, they can be entirely orphaned as in this example).

We'll need to be able to identify the scope they were created in (if it is still reachable) when we try to dump them, so record this information as part of the survey.