Factor out common code from Interpreter.p.step and .run
The code which actually performs a single step of the interpreter was repeated in Interpreter.prototype.step and Interpreter.prototype.run because it was believed that manually inlining it would improve performance.

It turns out this was false, and factoring it out into a separate method actually improves overall performance on the fibbonacci microbenchmark by about 5%!

This improvement also partially mitigates the slowdown caused by serialisation tests noted in bug #193.