Use single-quoted strings where possible
The testcases.js file originated in the code base for the initial golang implementaiton of the Code City JS interpreter.  That interpreter never had support for single-quoted strings, so all test input used double-quoted strings.  Since we no longer have that restriciton, rewrite the file to be slightly more styleguide-compliant.