Fix Error constructor wrong-prototype bug
Invoking an error constructor after deserialisation caused the resulting error objects to have an incorrect prototpye chain: the prototype would be the <Foo>Error.prototype object created by the (second) Interpreter constructor rather than the object deserialised from disk (and this object and its prototypes would have no properties since the startup .js files hadn't ever been run on that Interpreter).

The root cause of this was that the constructors were using the proto object captured in a closure at the time of the constructor's construction, rather than using intrp.ERROR etc.

The solution is to use intrp.ERROR et al.  This patch does that, albeit in a sub-optimal way, mixing usage of intrp.ERROR and intrp['ERROR'], which would break if we used JS Compiler output.  Since we only use the compiler for type-checking, we can get away with it.