Config file change: specify origin, not redirectUri
Change the config file to contain the (common) origin for the login and connect pages, and construct the redirect URI by combining origin + loginPage.

Note that you will need to update your loginServer.cfg file when applying this commit!  Just replace the line that looks like:

  "redirectUri": "https://example.codecity.world/login",

with one that looks like

  "origin": "https://example.codecity.world",