Script to start up nginx for local dev
The exisiting cc-dev.conf file requires that it be copied to /etc/nginx and edited, and nginx be started as root in order to have access to /var/run (or /usr/local/var/run, depending on how nginx was built).  If you have a nonstandard Homebrew set-up, even that may not be enough, because  the paths compiled into nginx may not exist.

So instead, bin/nginx-dev create a <repo>/var directory (if it doesn't already exist), write a custom config file to it, then start nginx in the foreground, without sudo, independently of any other nginx configuration running on the machine.