Fix autocomplete menus that are too long.
The magic number of 12 is made up of 8 pixels for the top margin of the body, plus 2*1 pixels for the top/bottom borders of the menu, plus 2 pixels to not actually hit the page bottom.
Choosing not to comment this, since our code isn’t being compiled, and CSS is full of magic numbers anyway.