Fix IterableWeakMap circular collection issue
Also solves layered collection issue.

The problem was that the (iterable) Set of Cells was creating strong references to all the values.  By instead keeping a Set of WeakRefs, the values are now only strongly referred to via the WeakMap, allowing them to be garbage collected as soon as their respective keys are unreachable.