Whitespace
Styleguide now clarifies that there should be a single space "separating any reserved word (such as if, for, or catch) *except for function and super*, from an open parenthesis (() that follows it on that line" (emphasis mine).

Also remove some extraneous spaces from some object literals and short function bodies.