Suppress fewer error messages
Make better use of the options parameter to the run*Test functions, to ensure that we only suppress log output that we are expecting.

This helps with debugging tests that are mysteriously failing (mysteriously because e.g. there is a syntax error in the code fed to the interpreter, but the resulting unhandled exception is suppressed).