Fix failing test of connectionListen
As of macOS 10.14 Mojave, ports <1024 are no longer privileged, so this test was failing because it was successfully able to bind port 22.

There's no obvious alternative candidate for a port that macOS will certainly be listening on, or will certainly refuse to let us bind, so just remove that test case entirely for now.