Fix pathological code RegExps
The RegExps for matching single- and double-quoted strings were inadvertently prone to exponential backtracking on strings that began with a quote character but did not terminate with one.

This is in aid of issue #325.

While I'm at it, make all the non-`exact` RegExps global, for consistency with regexps.singleQuotedString.