Improve editor’s butter
Move CSS to editor’s style sheet instead of global sheet for all of code.
Raise z-index so butter appears on top of tabs.
Print xhr http error messages (404, 500) in butter instead of console.
Remove mask from editor in the case of an xhr http error (404, 500) so that a resave may be attempted, or at least the code may be copied.