Make polyfill functions non-anonymous
We previouly had an issue with polyfill Function object .name
properties all being set to 'value'.

This was caused by creating these method with Object.defineProperty:
the object expression containing the property attrributes would set
the .name of the anonymous function expressions to 'value' because
they are assigned to the value property of the property descriptor
(and the ES6 mandates that anonymous function expressions assigned to
variables or appearing in object literals get a .name macthing the
variable/property they are assigned to).

This was mostly fixed in 6d5c4d7 (PR #382), when we directly
assigned the function expression to the property and then called
Object.defineProperty after the fact to make the new property
non-enumerable.

To be extra sure, however, this commit makes the function expressions
themselves non-anonymous.  This also makes the code here consistent
with our usual practice in the core.

(This change has been applied to the dogfood database, except for
Array.prototype.join which is defined in a closure and so not safe
to edit.)