Dumper.p.survey_: use Map instead of .preferredBadness
Since we only need to know the badness of the preferred reference while we are carrying out Dijkstra's algorithm, store this in a Map in Dumper.prototype.survey_ instead of on the .preferredBadness property of ObjectDumper instances.

Also fix an error in the algorithm that was failing to survey objects accessible only via local variables in non-global scopes.

Modify the tests to take into account both of the above.