testDumperSurvey: remove Dijkstra-related tests
Revert the changes made in b205562 to testDumperSurvey to add some tests for the implementation of Dijkstra's method in Dumper.prototype.survey_.

Instead, keep this test method focused on testing the SubDumper.prototype.survey methods (and update docs to reflect that).

Tests for Dumper.prototype.survey_ will be added separately.