Sprite the connect/reload icons.
Previously they were separate files.  Which meant that if the network failed, and the user was disconnected, the reload icon could not be loaded and would appear as a broken image.  Now it should be cached since it is the same file as the connect icon.

Also, one less HTTP request.