ObjectDummper.p.dump: dump extensibility correctly
Previously, ObjectDumper.prototype.dump would incorrectly prevent an object from being marked as DONE if treeOnly was true and some of the value objects were out-of-tree: those properties were (for the purposes of computing Done-ness) treated as if they were skipped, even though their value and attributes had both been set.  This prevented a call to Object.preventExtensions from being emitted.

After fixing this, it came to light that nothing prevented a second call to Object.preventExtensions being emitted if the object (having been DONE but not DONE_RECURSIVELY) was re-visited later.  This was fixed by checking to make sure we only emit the call the first time the ObjectDumper transitions from Done.NO to Done.DONE.