Dumper: pruneRest - mark objects not to be dumped recursively
It turns out we would like to be able to mark objects such that we can dump individual properties of them, but that if they are later encountered in a recursive dump no addiitonal properties will be dumped.  This is more akin to prune than skip, so the new feature is named "pruneRest".