Check in production core
There is still a lot of cruft to be cleaned up, but the files in core/ should be (modulo a call to $.system.onStartup() and moving a few object to $.startRoom) be sufficient to start a new server.

Not checking in any .js files in database/ (and removing the one that was there already) as non-core database material is (or will be) user data and doesn't belong in git.