Work around issue with #! line on Linux
Linux, when running a #! script, allows only one argument to be passed to the interpreter—so

    #!/usr/bin/env node --harmony-weak-refs

results in "/usr/bin/env: ‘node --harmony-weak-refs’: No such file or directory".

macOS spits arguments as usual, so no problem.  One possible fix is to use the -S option to env (which cause env to do the splitting), but this requies coreutils >= v8.30, and Debian 9 Stretch only has v8.26.

So, for now, just do a (very partial) version of the change to etc/codecity.service made in 3c451cf.