Correct cookie paths
According to the spec:

4.1.2.4. The Path Attribute
The scope of each cookie is limited to a set of paths, controlled by the Path attribute. If the server omits the Path attribute, the user agent will use the "directory" of the request-uri's path component as the default value.

So this means that by not specifying a path, the path of the ID cookie was implcity being set to ‘/login’ when running in single-domain mode, and was thus only sent to the login page.  Whereas in subdomain mode the path was being implicitly set to ‘/‘ and was thus sent everywhere.

What I fail to understand is why this worked previously.  This code was never changed, yet it seemed to work fine before.

This PR also edits the path for the client’s tab cookie, by removing the path so that it floats to the current URL.  This cookie was verified broken in prod (fixed now), meaning that the world/log preference is lost when reloading.