Make loginServer URL-independent
For Blockly Games, we want to be able to have the login server
respond to requests on http://blockly.games/login and
http://admin.blockly.games/login as well as various other staging
URLs.  To facilitate this, modify the loginServer to work out what
URL the client thinks it's talking to, and use that as the URL
to pass to the Oauth2 API.  This entails creating a separate
google.auth.Oauth2 client object for each such URL, and does not
obviate the need for each such URL to be listed in as an authorised
URL in the GCP Oauth console, but it does mean that a single
loginServer can handle an arbitrary set of login URLs with no
need for configuration.

Using https://www.npmjs.com/package/forwarded-parse to parse the
Forwarded header as an interim measure: we will most probably want
a similar parser in the core $.servers.http soon enough at which
point we can replace this dependency with our own implementation.