Fix repl output after .pseudoToNative changes
In PR #450 Interpreter.prototype.pseudoToNative was modified to make
the JSON.stringify builtin behave more correctly (per ES5.1 spec).

Unfortunately this meant that in repl any expression that yeilded
a function would be shown as having value undefined.

Fix repl so that if the expression evaluates to a function that
function is printed.  (It will still hide methods, as JSON.stringify
does.  But this is better than before, and perfecting repl is not
a priority at the moment.)