Interpreter.prototype.Server refactor
This change was motivated by the two tests which are added by this
commit.  The first, testServerPauseStart, was already passing without
issue, but the second was causing issues because the .listen and .close
operation on net.Server instances are actually asynchronous, and so the
calling .stop and then .start immediately on an Interpreter instance
was leaving some of the Interpreter.prototype.Server instances with
null .server_ properties incorrectly.

This is solved by creating a single net.Server instance in the
Interpreter.prototype.Server constructor, saving it in .server_, and
then taking advantage of the fact that a net.Server can have .listen
called more than once (if you call .close in between!).

This necessitated some more-extensive-than-expected changes to
Serializer to ensure that the initialised .server_ property would not
be overwritten by a dangling null value when deserialising.