Fix failing IterableWeakMap/Set tests
Somewhere between Node.js v12.12.0 and v12.22.1 the implementation
of the interaction between the event loop and FinalizationGroup (now
FinalizationRegistry) was changed slightly.  We now need to be a little
more fastidious about making sure that the event loop is fully cycled
to ensure that finalisers are run promptly if we want to test
finalisation behaviour.

(As ever there is actually no guarantee that finalisers are ever run,
but in fact V8 is pretty reliable about doing so provided on has jumped
through all the spec-mandated hoops.)