Fix incorrect handling of primitives in JSON.parse
When JSON.parse is given null or undefined it should throw SyntaxError
but, due to attempting to call .toString() on the argument value
it was incorrectly throwing TypeError instead.

Add tests for these and various other edge cases.

Also add tests for similar edge cases in JSON.stringify, as well as
improving the names of a few existing tests.