Rewrite RegExp constructor as a NativeFunction
The goal was to remove the .toString calls in the old implementation,
but it was rewritten to be (much more) spec-compliant.

The new implementation is based on a mixture of the ES5.1, ES6 and
ES2022 specifications: we disallow flags not permitted in ES5.1 (for
consistency with Acorn's enforcement of that rule for RegExp literals),
but otherwise generally follow the ES6/ES2022 behaviour (e.g., we
permit supplying both a RegExp-valued pattern argument and a
non-undefined flags argument).

Also remove two additional uses of .toString() in RegExp.prototype
methods that could cause the interpreter to crash.

Fixes #474.