Trust Acorn to check parameter legality in builtin Function constructor
Since Acorn is going to verify that the parameter list in the assembled
function expression is valid anyway, trust it to do that validation
and don't try to pre-guess it.  This allows us to use non-ASCII
unicode letters in function parameter names created via the Function
constructor just as can already be done with function expressions.