Modify boilerplate used by Function constructor
This is change 2 of 2(?) to implement useful clarifications of the
behaviour of Function.prototype.toString adopted in ES2019 per the
"Function.prototype.toString Revision" proposal
(https://tc39.es/Function-prototype-toString-revision/).

In particular, per
https://tc39.es/Function-prototype-toString-revision/#proposal-sec-createdynamicfunction
the functions returned by the Function constructor are always named
"anonymous" (and this appears in their .toString() output), and have
three extra newline characters to ensure that they will be syntactically
valid even if arguments to the Function constructor contains trailing
line comments.